refactor(e2e): dissolve mbps tests into consensus and prune suites
Dissolves the three MBPS test files in the multi-node category, relocating
every unique assertion verbatim:

- consensus/mbps.parallel.test.ts: the proposed-anchor monotonicity, L2->L1,
  L1->L2, non-validator re-exec/cold-sync, and deploy+call sub-slot ordering
  its from mbps.parallel; the proposer-pipelining offset + blob-promotion it
  from mbps.pipeline.parallel; and both redistribution its (budget
  redistribution kept verbatim, multiplier asymmetry) from mbps_redistribution.
- prune/pipeline_prune.parallel.test.ts: the prune-on-skip-publish-under-
  pipelining it from mbps.pipeline.parallel.

mbps.parallel it 1 (checkpointed-anchored MBPS block-count + proven) is dropped
as redundant: the block-count and proven-checkpoint properties are already
covered by consensus/block_building/high_tps.

assertProposerPipelining is lifted onto SingleNodeTestContext (with a shared
BlockProposedEvent type) so both relocated pipelining its reuse it. test_patterns
flake entries for the prune and redistribution its move to their new paths.

This is a behavior-preserving relocation, revertible as a unit.
