+++ .test_patterns.yml
@@ -271,14 +271,14 @@ tests:
     owners:
       - *phil
 
-  - regex: "mbps.pipeline.parallel.test.ts.*prunes uncheckpointed"
+  - regex: "src/multi-node/prune/pipeline_prune.parallel.test.ts.*prunes uncheckpointed"
     owners:
       - *sean
 
   # Asserts all late txs land in the checkpoint's last block, but the proposer seals the last block as
   # soon as one tx is available and snapshots the mempool, so late txs still propagating over gossip
   # spill into the next checkpoint. Flaky until the sequencer waits for the last block's tx window.
-  - regex: "src/multi-node/mbps_redistribution.test.ts"
+  - regex: "src/multi-node/consensus/mbps.parallel.test.ts.*redistributes checkpoint budget"
     owners:
       - *phil
