+++ yarn-project/end-to-end/src/multi-node/prune/pipeline_prune.parallel.test.ts
@@ -1,32 +1,31 @@
 import type { Archiver } from '@aztec/archiver';
 import type { AztecNodeService } from '@aztec/aztec-node';
-import { AztecAddress, EthAddress } from '@aztec/aztec.js/addresses';
+import type { AztecAddress, EthAddress } from '@aztec/aztec.js/addresses';
 import { NO_WAIT } from '@aztec/aztec.js/contracts';
 import { Fr } from '@aztec/aztec.js/fields';
 import type { Logger } from '@aztec/aztec.js/log';
 import { waitForTx } from '@aztec/aztec.js/node';
 import type { EpochCacheInterface } from '@aztec/epoch-cache';
-import { RollupContract } from '@aztec/ethereum/contracts';
-import type { Operator } from '@aztec/ethereum/deploy-aztec-l1-contracts';
 import { asyncMap } from '@aztec/foundation/async-map';
 import { BlockNumber, CheckpointNumber, SlotNumber } from '@aztec/foundation/branded-types';
-import { times, timesAsync } from '@aztec/foundation/collection';
-import { SecretValue } from '@aztec/foundation/config';
+import { timesAsync } from '@aztec/foundation/collection';
 import { retryUntil } from '@aztec/foundation/retry';
-import { bufferToHex } from '@aztec/foundation/string';
 import { executeTimeout } from '@aztec/foundation/timer';
-import { TestContract } from '@aztec/noir-test-contracts.js/Test';
+import type { TestContract } from '@aztec/noir-test-contracts.js/Test';
 import type { SequencerEvents } from '@aztec/sequencer-client';
 import { L2BlockSourceEvents } from '@aztec/stdlib/block';
-import { getTimestampForSlot } from '@aztec/stdlib/epoch-helpers';
 
 import { jest } from '@jest/globals';
-import { privateKeyToAccount } from 'viem/accounts';
 
-import { type EndToEndContext, getPrivateKeyFromIndex } from '../fixtures/utils.js';
-import { TestWallet } from '../test-wallet/test_wallet.js';
-import { proveInteraction } from '../test-wallet/utils.js';
-import { MultiNodeTestContext } from './multi_node_test_context.js';
+import type { EndToEndContext } from '../../fixtures/utils.js';
+import type { TestWallet } from '../../test-wallet/test_wallet.js';
+import { proveInteraction } from '../../test-wallet/utils.js';
+import {
+  type BlockProposedEvent,
+  MultiNodeTestContext,
+  type RegisteredValidator,
+  buildMockGossipValidators,
+} from '../multi_node_test_context.js';
 
 jest.setTimeout(1000 * 60 * 20);
 
@@ -37,24 +36,23 @@ const EXPECTED_BLOCKS_PER_CHECKPOINT = 8;
 const TX_COUNT = 34;
 
 /**
- * E2E tests for proposer pipelining with Multiple Blocks Per Slot (MBPS).
- * Verifies that the block proposer in slot N is the validator scheduled on L1 for slot N+1
- * (the proposer view uses a +1 slot offset).
+ * E2E prune-and-recover test under proposer pipelining with MBPS. A selected next proposer is configured
+ * to skip its checkpoint publish mid-run, which triggers an uncheckpointed-blocks prune; publishing is
+ * then re-enabled and the chain recovers. Asserts that recovery still produces a multi-block checkpoint
+ * with the correct pipelining build-vs-submission slot offset, and that the recovered block number is
+ * past the pre-prune baseline.
  *
- * Four-validator suite with a prover node (fake proofs) and 500ms mock gossip latency to simulate
- * adverse network conditions. Two tests: (1) normal pipelining flow asserting build-vs-submission
- * slot offsets and blob-fetch promotion; (2) a proposer skips its checkpoint publish, triggering an
- * uncheckpointed-blocks prune followed by recovery. Uses MultiNodeTestContext with mockGossipSubNetwork
- * and no initial sequencer.
+ * Four-validator suite with a prover node (fake proofs) and 500ms mock gossip latency to simulate adverse
+ * network conditions. Relocated from the dissolved `mbps.pipeline.parallel` file. Uses MultiNodeTestContext
+ * with mockGossipSubNetwork and no initial sequencer.
  */
-describe('multi-node/mbps_pipeline', () => {
+describe('multi-node/prune/pipeline_prune', () => {
   let context: EndToEndContext;
   let logger: Logger;
-  let rollup: RollupContract;
   let archiver: Archiver;
 
   let test: MultiNodeTestContext;
-  let validators: (Operator & { privateKey: `0x${string}` })[];
+  let validators: RegisteredValidator[];
   let nodes: AztecNodeService[];
   let contract: TestContract;
   let wallet: TestWallet;
@@ -68,11 +66,7 @@ describe('multi-node/mbps_pipeline', () => {
   }) {
     const { syncChainTip = 'checkpointed', ...setupOpts } = opts;
 
-    validators = times(NODE_COUNT, i => {
-      const privateKey = bufferToHex(getPrivateKeyFromIndex(i + 3)!);
-      const attester = EthAddress.fromString(privateKeyToAccount(privateKey).address);
-      return { attester, withdrawer: attester, privateKey, bn254SecretKey: new SecretValue(Fr.random().toBigInt()) };
-    });
+    validators = buildMockGossipValidators(NODE_COUNT);
 
     test = await MultiNodeTestContext.setup({
       numberOfAccounts: 0,
@@ -93,8 +87,8 @@ describe('multi-node/mbps_pipeline', () => {
       skipInitialSequencer: true,
     });
 
-    ({ context, logger, rollup } = test);
-    wallet = context.wallet;
+    ({ context, logger } = test);
+    wallet = context.wallet as TestWallet;
     from = context.accounts[0]; // auto-created by setup
 
     logger.warn(`Initial setup complete. Starting ${NODE_COUNT} validator nodes.`);
@@ -166,150 +160,9 @@ describe('multi-node/mbps_pipeline', () => {
     return multiBlockCheckpointNumber!;
   }
 
-  /** Waits until a specific multi-block checkpoint is proven. */
-  async function waitForProvenCheckpoint(targetCheckpoint: CheckpointNumber) {
-    const provenTimeout = test.L2_SLOT_DURATION_IN_S * test.epochDuration * 4;
-    logger.warn(`Waiting for checkpoint ${targetCheckpoint} to be proven (timeout=${provenTimeout}s)`);
-    await test.waitUntilProvenCheckpointNumber(targetCheckpoint, provenTimeout);
-    logger.warn(`Proven checkpoint advanced to ${test.monitor.provenCheckpointNumber}`);
-  }
-
-  /**
-   * Asserts pipelining by comparing the build slot (from block-proposed events) against
-   * the submission slot (from block headers). With pipelining, the block is built in slot N
-   * but its header carries submission slot N+1.
-   */
-  async function assertProposerPipelining(
-    blockProposedEvents: { blockNumber: BlockNumber; slot: SlotNumber; buildSlot: SlotNumber }[],
-    logger: Logger,
-  ) {
-    const checkpoints = await archiver.getCheckpoints({ from: CheckpointNumber(1), limit: 50 });
-    const allBlocks = checkpoints.flatMap(pc => pc.checkpoint.blocks);
-
-    logger.warn(`assertProposerPipelining: ${allBlocks.length} blocks, ${blockProposedEvents.length} events`, {
-      blockNumbers: allBlocks.map(b => b.number),
-      eventBlockNumbers: blockProposedEvents.map(e => e.blockNumber),
-    });
-
-    let foundPipelining = false;
-
-    for (const block of allBlocks) {
-      const headerSlot = block.header.globalVariables.slotNumber; // submission slot (N+1)
-      const coinbase = block.header.globalVariables.coinbase;
-
-      // Find the block-proposed event for this block (use Number() for safe comparison)
-      const event = blockProposedEvents.find(e => Number(e.blockNumber) === Number(block.number));
-      // if there is no event, then it was probably block number one - which was proposed in setup
-      if (!event) {
-        continue;
-      }
-
-      const buildSlot = event.buildSlot; // build slot (N)
-
-      // Verify the pipelining offset: block built in slot N, submitted in slot N+1
-      expect(Number(headerSlot)).toBe(Number(buildSlot) + 1);
-      foundPipelining = true;
-
-      // Verify coinbase matches the expected proposer for the submission slot
-      const expectedProposer = await rollup.getProposerAt(getTimestampForSlot(headerSlot, test.constants));
-      expect(coinbase).toEqual(expectedProposer);
-
-      logger.warn(`Block ${block.number}: buildSlot=${buildSlot}, submissionSlot=${headerSlot}, coinbase=${coinbase}`, {
-        blockNumber: block.number,
-        buildSlot,
-        headerSlot,
-        coinbase: coinbase.toString(),
-        expectedProposer: expectedProposer.toString(),
-      });
-    }
-
-    expect(foundPipelining).toBe(true);
-    logger.warn(`Pipelining assertion passed for ${allBlocks.length} blocks`);
-  }
-
   afterEach(async () => {
     jest.restoreAllMocks();
     await test?.teardown();
-  });
-
-  // Pre-proves TX_COUNT txs, starts sequencers, waits for all txs to be mined. Asserts a
-  // MBPS checkpoint with ≥EXPECTED_BLOCKS_PER_CHECKPOINT blocks. Asserts every block's header
-  // slot equals build-slot+1 (pipelining offset). Verifies node-0 fetches blobs (promotion
-  // disabled) while nodes 1-3 skip blob fetching (promotion enabled). Waits for the checkpoint
-  // to be proven.
-  it('pipelining builds blocks using slot plus 1 proposer and proves them', async () => {
-    await setupTest({ syncChainTip: 'checkpointed', minTxsPerBlock: 1, maxTxsPerBlock: 2 });
-
-    // Spy on getBlobSidecar on all validator nodes before sequencers start, so we check that nodes
-    // promote their proposed checkpoints and don't source data from blobs if they don't need to.
-    const blobSpies = nodes.map((node, i) => {
-      const blobClient = node.getBlobClient()!;
-      const spy = jest.spyOn(blobClient, 'getBlobSidecar');
-      logger.warn(`Installed getBlobSidecar spy on validator node ${i}`);
-      return spy;
-    });
-
-    // Subscribe to block-proposed events to capture build slots
-    const blockProposedEvents: { blockNumber: BlockNumber; slot: SlotNumber; buildSlot: SlotNumber }[] = [];
-    const sequencers = nodes.map(n => n.getSequencer()!);
-    for (const sequencer of sequencers) {
-      sequencer.getSequencer().on('block-proposed', (args: Parameters<SequencerEvents['block-proposed']>[0]) => {
-        logger.warn(`block-proposed event: blockNumber=${args.blockNumber}, slot=${args.slot}`, args);
-        blockProposedEvents.push({
-          blockNumber: args.blockNumber,
-          slot: args.slot,
-          buildSlot: args.buildSlot,
-        });
-      });
-    }
-
-    const initialCheckpointNumber = await rollup.getCheckpointNumber();
-    logger.warn(`Initial checkpoint number: ${initialCheckpointNumber}`);
-
-    // Pre-prove and send transactions
-    const txs = await timesAsync(TX_COUNT, i =>
-      proveInteraction(context.wallet, contract.methods.emit_nullifier(new Fr(i + 1)), { from }),
-    );
-    const txHashes = await Promise.all(txs.map(tx => tx.send({ wait: NO_WAIT })));
-    logger.warn(`Sent ${txHashes.length} transactions`, { txs: txHashes });
-
-    // Start the sequencers
-    await Promise.all(sequencers.map(s => s.start()));
-    logger.warn(`Started all sequencers`);
-
-    // Wait until all txs are mined
-    const timeout = test.L2_SLOT_DURATION_IN_S * 5;
-    const receipts = await executeTimeout(
-      () => Promise.all(txHashes.map(txHash => waitForTx(context.aztecNode, txHash, { timeout }))),
-      timeout * 1000,
-    );
-    logger.warn(`All txs have been mined`);
-
-    // Verify MBPS works with pipelining; target the highest block number across mined receipts
-    const maxMinedBlockNumber = BlockNumber(Math.max(...receipts.map(r => r.blockNumber ?? 0)));
-    const multiBlockCheckpoint = await assertMultipleBlocksPerSlot(
-      EXPECTED_BLOCKS_PER_CHECKPOINT,
-      maxMinedBlockNumber,
-      logger,
-    );
-
-    // Verify the pipelining offset: build slot N vs submission slot N+1
-    await assertProposerPipelining(blockProposedEvents, logger);
-
-    // Verify blob fetching behavior: node 0 has promotion disabled so it must fetch blobs,
-    // while all other nodes should promote their proposed checkpoints and skip blob fetching entirely.
-    for (let i = 0; i < blobSpies.length; i++) {
-      const calls = blobSpies[i].mock.calls.length;
-      logger.warn(`Validator ${i} made ${calls} getBlobSidecar calls`);
-      if (i === 0) {
-        expect(calls).toBeGreaterThan(0);
-      } else {
-        expect(calls).toBe(0);
-      }
-    }
-
-    // Verify proving still works end-to-end with pipelined proposers
-    await waitForProvenCheckpoint(multiBlockCheckpoint);
   });
 
   // Establishes a baseline at checkpoint 1. Identifies the next proposer and disables its
@@ -319,7 +172,7 @@ describe('multi-node/mbps_pipeline', () => {
   it('prunes uncheckpointed blocks when proposer fails to deliver', async () => {
     await setupTest({ syncChainTip: 'checkpointed', minTxsPerBlock: 1, maxTxsPerBlock: 2 });
 
-    const blockProposedEvents: { blockNumber: BlockNumber; slot: SlotNumber; buildSlot: SlotNumber }[] = [];
+    const blockProposedEvents: BlockProposedEvent[] = [];
     const sequencers = nodes.map(n => n.getSequencer()!);
 
     // Pre-prove and send transactions
@@ -394,7 +247,7 @@ describe('multi-node/mbps_pipeline', () => {
     await assertMultipleBlocksPerSlot(EXPECTED_BLOCKS_PER_CHECKPOINT, maxMinedBlockNumber, logger);
 
     // Verify the pipelining offset: build slot N vs submission slot N+1
-    await assertProposerPipelining(blockProposedEvents, logger);
+    await test.assertProposerPipelining(archiver, blockProposedEvents, logger);
 
     const recoveredBlockNumber = await archiver.getBlockNumber();
     logger.warn(`Recovery complete: block number ${recoveredBlockNumber} > ${checkpointedBlockNumber}`);
